Added select method to the database class and updated the userList to use it. See CHANGELOG.md

+++ views/user/index.php
@@ -23,22 +23,27 @@
         <th>Edit</th>
         <th>Delete</th>
         <?php
-        foreach ($this->userList as $key => $value) {
-            echo '<tr>'; //table row
-            echo '<td>' . $value['id'] . '</td>';
-            echo '<td>' . $value['login'] . '</td>';
-            echo '<td>' . $value['role'] . '</td>';
-            echo '<td> &nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</td>';
-            if ($value['role'] === 'owner') { //custom code, it hides the delete button for owners. This is more for aesthetics than anything; the delete function is set to ignore owners
-                echo '<td>
+        if (!empty($this->userList)) {
+            foreach ($this->userList as $key => $value) {
+                echo '<tr>'; //table row
+                echo '<td>' . $value['id'] . '</td>';
+                echo '<td>' . $value['login'] . '</td>';
+                echo '<td>' . $value['role'] . '</td>';
+                echo '<td> &nbsp;&nbsp;&nbsp;|&nbsp;&nbsp;&nbsp;</td>';
+                if ($value['role'] === 'owner') { //custom code, it hides the delete button for owners. This is more for aesthetics than anything; the delete function is set to ignore owners
+                    echo '<td>
                     <a href="' . URL . 'user/edit/' . $value['id'] . '">Edit</a></td>';
-            } else {
-                echo '<td>
+                } else {
+                    echo '<td>
                     <a href="' . URL . 'user/edit/' . $value['id'] . '">Edit</a>&nbsp;</td>
                     <td><a href="' . URL . 'user/delete/' . $value['id'] . '">Delete</a>' .  '</td>';
+                }
+                echo '</tr>';
             }
-            echo '</tr>';
+        } else {
+            echo '<h2 style="color:red;">No Users found!</h2>';
         }
+
         // print_r($this->userList);
         ?>
     </tr>
